requirements/dev.txt
# Local development dependencies go here
-r base.txt

# Coverage
# ------------------------------------------------
coverage==4.2
django-coverage-plugin==1.3.1

# Fixtures
# ------------------------------------------------
factory_boy==2.11.1

# Utils
# ------------------------------------------------
django-extensions==1.7.4
ipython==5.1.0

# Debug
# ------------------------------------------------
ipdb==0.10.1
django-debug-toolbar==1.6

# Mock
# ------------------------------------------------
requests-mock==1.5.2
